Split and improve tests
